Revert "Navigationbar"
