+++ src/components/profile-view/profile-view.jsx
@@ -1,7 +1,7 @@
 import { UserInfo } from "./user-info";
 import { UserEdit } from "./user-edit";
 import { MovieCard } from "../movie-card/movie-card";
-import { Col, Button, Row } from "react-bootstrap";
+import { Col, Button } from "react-bootstrap";
 
 export const ProfileView = ({
   user,
@@ -14,31 +14,54 @@ export const ProfileView = ({
     user.FavoriteMovies.includes(movie.id)
   );
 
+  const deleteAccount = () => {
+    fetch(`https://movie-api-movieflix.onrender.com/users/${user.Username}`, {
+      method: "DELETE",
+      headers: { Authorization: `Bearer ${token}` },
+    })
+      .then((response) => {
+        if (response.ok) {
+          alert("Account has been deleted!");
+          onLoggedOut();
+        } else {
+          alert(" Could not delete account");
+        }
+      })
+      .catch((e) => {
+        alert(e);
+      });
+  };
   return (
     <>
-      <Row className="d-flex justify-content-center">
-        <Col xl={6} md={12} className="d-flex justify-content-center mh-60">
-          <UserInfo user={user} />
+      <Col>
+        <UserInfo user={user} />
+        <UserEdit
+          user={user}
+          token={token}
+          updateUser={updateUser}
+          onLoggedOut={onLoggedOut}
+        />
+        <Button
+          className="mt-4"
+          variant="danger"
+          type="submit"
+          onClick={() => {
+            if (confirm("Are you sure you want to delete account?")) {
+              deleteAccount();
+            }
+          }}
+        >
+          Delete your Account
+        </Button>
+      </Col>
+      <Col>
+        <h3 className="mt-3 mb-3 text-dark"> Your favorite movies :</h3>
+      </Col>
+      {favoriteMovies.map((movie) => (
+        <Col key={movie.id} className="mb-4" xl={2} lg={3} md={4} xs={6}>
+          <MovieCard movie={movie} />
         </Col>
-        <Col xl={6} md={12} className="d-flex justify-content-center mh-60">
-          <UserEdit
-            user={user}
-            token={token}
-            updateUser={updateUser}
-            onLoggedOut={onLoggedOut}
-          />
-        </Col>
-        <Col>
-          <h3 className="mt-3 mb-3 text-dark"> Your favorite movies :</h3>
-        </Col>
-      </Row>
-      <Row className="d-flex ">
-        {favoriteMovies.map((movie) => (
-          <Col key={movie.id} className="mb-4 mt-4" xl={6} md={12}>
-            <MovieCard movie={movie} />
-          </Col>
-        ))}
-      </Row>
+      ))}
     </>
   );
 };
